[BERDA] Changed settings page.
