+++ examscheduler.client/src/Components/ScrollableContainer/ScrollableContainer.css
@@ -4,7 +4,6 @@
     flex-direction: column;
     gap: 20px;
     overflow-y: auto;
-    /* max-height: 80vh; */
     padding: 20px;
     box-sizing: border-box;
     background-color: #f5f5f5;
