+++ examscheduler.client/src/Components/StylizedBlock/StylizedBlock.css
@@ -3,10 +3,8 @@
     background-color: #ffffff;
     border-radius: 8px;
     box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-    /*margin: 20px 0;*/
     padding: 20px;
     width: 100%;
-    /* max-width: 600px; */
     box-sizing: border-box;
     transition: transform 0.2s, box-shadow 0.2s;
 }
