+++ examscheduler.client/src/Forms/ChangePasswordForm.css
@@ -1,0 +1,107 @@
+/* Form container */
+.change-password-form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 5px;
+}
+
+    /* Form header */
+    .change-password-form h2 {
+        text-align: center;
+        margin-bottom: 20px;
+    }
+
+    /* Form group container */
+    .change-password-form div {
+        display: flex;
+        align-items: center;
+        /* justify-content: center; */
+        width: 60%;
+        /* max-width: 400px; */
+        gap: 10px;
+        /* align-content: space-between; */
+        /* flex-wrap: nowrap; */
+        flex-direction: row;
+    }
+
+    /* Form labels */
+    .change-password-form label {
+        font-size: 1rem;
+        font-weight: 500;
+        color: #444;
+        width: 35%; /* Adjust label width as needed */
+        text-align: left;
+        margin-right: 10px;
+    }
+
+    /* Form inputs */
+    .change-password-form input {
+        flex: 1;
+        padding: 10px 15px;
+        font-size: 1rem;
+        border: 1px solid #ccc;
+        border-radius: 4px;
+        box-sizing: border-box;
+        transition: border-color 0.3s ease;
+        width: 100%;
+    }
+
+        /* Input focus state */
+        .change-password-form input:focus {
+            border-color: #007bff;
+            outline: none;
+            box-shadow: 0px 0px 3px rgba(0, 123, 255, 0.5);
+        }
+
+    /* Submit button */
+    .change-password-form button {
+        padding: 10px 20px;
+        font-size: 1rem;
+        font-weight: 600;
+        color: #fff;
+        background-color: #007bff;
+        border: none;
+        border-radius: 4px;
+        cursor: pointer;
+        transition: background-color 0.3s ease, transform 0.2s ease;
+        margin: 10px 10px 0px 0px;
+    }
+
+        /* Button hover state */
+        .change-password-form button:hover {
+            background-color: #0056b3;
+            transform: translateY(-1px);
+        }
+
+        /* Button active state */
+        .change-password-form button:active {
+            background-color: #004085;
+            transform: translateY(0px);
+        }
+
+/* Success and error messages */
+.alert {
+    padding: 10px 15px;
+    border-radius: 4px;
+    font-size: 0.9rem;
+    margin-bottom: 15px;
+    display: inline-block;
+    width: 100%;
+    max-width: 400px;
+    box-sizing: border-box;
+}
+
+/* Success message */
+.alert-success {
+    background-color: #d4edda;
+    color: #155724;
+    border: 1px solid #c3e6cb;
+}
+
+/* Error message */
+.alert-danger {
+    background-color: #f8d7da;
+    color: #721c24;
+    border: 1px solid #f5c6cb;
+}
